+++ packages/core/src/utils/custom-jwt/client-error.ts
