+++ packages/core/src/utils/custom-jwt/index.ts
@@ -1,3 +1,3 @@
 export * from './custom-jwt.js';
 export * from './types.js';
-export * from './cloud-client-response-error.js';
+export * from './client-error.js';
